# perf: reduce factory memory usage