- Previously we use concret values for fields in message struct, using `unsafe.Sizeof(mesgs)`, it will produce 3362000 B (3.36 MB)
```go
type message struct {
	Num    typedef.MesgNum
	Fields [256]proto.Field
}

var mesgs = [...]message{ /* data */ }

```

- We change it to use pointer values since most of the message only has small amount of fields, it will only produce 842960 (0,84 MB)
```go
type message struct {
	Num    typedef.MesgNum
	Fields [256]*proto.Field // <---- this
}
```

There is no significant performance reduction on decoding process after factory uses pointer values
```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/decoder
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
         │   new.txt   │           new2.txt            │
         │   sec/op    │   sec/op     vs base          │
Decode-4   108.6m ± 1%   108.8m ± 1%  ~ (p=0.529 n=10)

         │   new.txt    │              new2.txt               │
         │     B/op     │     B/op      vs base               │
Decode-4   67.39Mi ± 0%   67.39Mi ± 0%  -0.00% (p=0.050 n=10)

         │   new.txt   │            new2.txt             │
         │  allocs/op  │  allocs/op   vs base            │
Decode-4   900.0k ± 0%   900.0k ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

```

- Another change is proto messages variable used by CreateMesg is now using array to ensure O(1) lookup, it will only use 22960 (22.96 KB) on its empty state.